+++ common/src/main/java/com/avp/common/model/inventory/AVPInventoryBearer.java
@@ -1,6 +1,0 @@
-package com.avp.common.model.inventory;
-
-public interface AVPInventoryBearer {
-
-    AVPInventory getInventory();
-}